Move version to within each module
