+++ windows-agent/cmd/ubuntu-pro-agent/main.go
@@ -14,6 +14,7 @@ import (
 	"github.com/canonical/ubuntu-pro-for-wsl/common"
 	"github.com/canonical/ubuntu-pro-for-wsl/common/i18n"
 	"github.com/canonical/ubuntu-pro-for-wsl/windows-agent/cmd/ubuntu-pro-agent/agent"
+	"github.com/canonical/ubuntu-pro-for-wsl/windows-agent/internal/consts"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -76,7 +77,7 @@ func setLoggerOutput(a app) (func(), error) {
 	log.SetOutput(w)
 
 	fmt.Fprintf(f, "\n======= STARTUP =======\n")
-	log.Infof("Version: %s", common.Version)
+	log.Infof("Version: %s", consts.Version)
 
 	return func() { _ = f.Close() }, nil
 }
