+++ windows-agent/internal/consts/version.go
@@ -1,4 +1,4 @@
-package common
+package consts
 
 // Version is the version name for Ubuntu Pro for WSL.
 var Version = "Dev"
