+++ wsl-pro-service/internal/consts/version.go
@@ -1,0 +1,4 @@
+package consts
+
+// Version is the version name for Ubuntu Pro for WSL.
+var Version = "Dev"
